add safety check for total verificaiton

+++ statement_to_client/main.py
@@ -21,13 +21,13 @@ else:
 
 sent_to_numbers = []
 
-def get_receipts(bookingid:str):
+def get_receipts(unitname:str):
     if "statement2.xlsx" not in os.listdir("."):
         return []
     rs= []
     wb2 = openpyxl.open("statement2.xlsx")
     for row in wb2.active.rows:
-        if bookingid in row[7].value:
+        if unitname == row[8].value:
             r = {
                 "vr_no": row[3].value,
                 "amount": row[14].value,
@@ -57,11 +57,14 @@ for row in wb.active.rows:
     bookingdate = row[10].value.split("T")[0]
     output_file_name = f"file"
     print(bookingid)
-    reciepts = get_receipts(bookingid)
+    reciepts = get_receipts(unitname)
+    rtotal = sum([float(r["amount"]) for r in reciepts])
     rtitles = "Receipts"
+    """
     if len(reciepts) > 15:
         rtitles = ""
-        reciepts = []
+        reciepts = []"""
+    
     
     with app.app_context():
         template = open("templates/print_receipt.html").read()
@@ -118,6 +121,10 @@ xxx"""
     with open("done.json", "w") as f:
         json.dump(done, f, indent=4)
     # remove this file after sending
+    if rtotal != netadjusted:
+        print("Total not matching")
+        print(rtotal, netadjusted)
+        break
     os.remove(f"{output_file_name}.pdf")
     os.remove(f"{output_file_name}.html")
     os.remove(f"{output_file_name}.png")
